feat(eks): bottlerocket versoin follows the cluster k8s versoin (#10189)
This PR will implicitly select correct bottlerocket variant in consist with the  cluster k8s version. For example, if the cluster k8s version is `1.17`, previously only bottlerocket `aws-k8s-1.15` variant is available. Now the variant will be `aws-k8s-1.17`

BREAKING CHANGE: Clusters previously running k8s version other than `1.15` and bottlerocket AMI(`aws-k8s-1.15` variant) will trigger AMI and node replacement.

Closes: #10188 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
